Backfill Card Face Images during set ingestion
Now, when a set is ingested, the optional `?faces=y` query string
parameter can be set to force an S3 bucket refresh from scryfall image
sources.  This is an intelligent backfill; if the image is already in
our S3 bucket, there's no point in re-uploading it.

The `vcb` helper script also now supports the `--images` flag to set
ingestion, to trigger this behavior.

Fixes #11